use size_t instead of uint64_t for memory size

+++ src/siphash.c
@@ -299,7 +299,7 @@ do {					\
 } while (0)
 
 uint64_t
-sip_hash24(uint8_t key[16], uint8_t *data, uint64_t len)
+sip_hash24(uint8_t key[16], uint8_t *data, size_t len)
 {
     uint64_t k0, k1;
     uint64_t v0, v1, v2, v3;
